+++ packages/data-model/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mlightcad/data-model
 
+## 1.2.2
+
+### Patch Changes
+
+- feat: add properties 'extmin' and 'extmax' for class AcDbDatabase
+
 ## 1.2.1
 
 ### Patch Changes